requirements.txt
ibm_watson
pylint
pytest
python-decouple
